feat: simplify provider setup of the app
