+++ services/orchest-webserver/client/src/App.tsx
@@ -5,16 +5,13 @@ import OpenReplay from "@openreplay/tracker";
 import { makeRequest } from "@orchest/lib-utils";
 import { enableMapSet } from "immer";
 import React from "react";
-import { BrowserRouter as Router, Prompt } from "react-router-dom";
+import { Prompt } from "react-router-dom";
 import { useIntercom } from "react-use-intercom";
 import { CommandPalette } from "./components/CommandPalette";
 import { OnboardingDialog } from "./components/layout/legacy/OnboardingDialog";
 import { SystemDialog } from "./components/SystemDialog";
-import { AppContextProvider } from "./contexts/AppContext";
 import { useGlobalContext } from "./contexts/GlobalContext";
-import { OrchestProvider } from "./contexts/Providers";
 import { HeaderBar } from "./header-bar/HeaderBar";
-import { useEditJob } from "./jobs-view/stores/useEditJob";
 import Jupyter from "./jupyter/Jupyter";
 
 enableMapSet();
@@ -27,13 +24,9 @@ const App = () => {
   // load server side config populated by flask template
   const {
     state: { hasUnsavedChanges },
-    setAsSaved,
     config,
     user_config,
   } = useGlobalContext();
-  const discardActiveCronJobChanges = useEditJob(
-    (state) => state.discardActiveCronJobChanges
-  );
 
   const jupyterRef = React.useRef<HTMLDivElement>(null);
 
@@ -81,56 +74,32 @@ const App = () => {
   };
 
   return (
-    <Router
-      getUserConfirmation={(message, callback) => {
-        // use Prompt component to intercept route changes
-        // handle the blocking event here
-        if (message === "hasUnsavedChanges") {
-          setConfirm(
-            "Warning",
-            "There are unsaved changes. Are you sure you want to navigate away?",
-            async (resolve) => {
-              setAsSaved();
-              discardActiveCronJobChanges();
-              callback(true);
-              resolve(true);
-              return true;
-            }
-          );
-        }
-      }}
-    >
-      <OrchestProvider>
-        <AppContextProvider>
-          <Box sx={{ display: "flex", flex: 1, overflow: "hidden" }}>
-            <HeaderBar />
-            <Box
-              component="main"
-              sx={{
-                flex: 1,
-                overflow: "hidden",
-                position: "relative",
-                minHeight: 0,
-                display: "flex",
-                flexDirection: "column",
-              }}
-              id="main-content"
-              data-test-id="app"
-            >
-              <Routes />
-              <div
-                ref={jupyterRef}
-                className="persistent-view jupyter hidden"
-              />
-            </Box>
-          </Box>
-          <Prompt when={hasUnsavedChanges} message="hasUnsavedChanges" />
-          <SystemDialog />
-          <OnboardingDialog />
-          <CommandPalette />
-        </AppContextProvider>
-      </OrchestProvider>
-    </Router>
+    <>
+      <Box sx={{ display: "flex", flex: 1, overflow: "hidden" }}>
+        <HeaderBar />
+        <Box
+          component="main"
+          sx={{
+            flex: 1,
+            overflow: "hidden",
+            position: "relative",
+            minHeight: 0,
+            display: "flex",
+            flexDirection: "column",
+          }}
+          id="main-content"
+          data-test-id="app"
+        >
+          <Routes />
+          <div ref={jupyterRef} className="persistent-view jupyter hidden" />
+        </Box>
+      </Box>
+
+      <Prompt when={hasUnsavedChanges} message="hasUnsavedChanges" />
+      <SystemDialog />
+      <OnboardingDialog />
+      <CommandPalette />
+    </>
   );
 };
