+++ services/orchest-webserver/client/src/contexts/Providers.tsx
@@ -1,9 +1,12 @@
+import { useEditJob } from "@/jobs-view/stores/useEditJob";
 import theme from "@/theme";
 import { ThemeProvider } from "@mui/material/styles";
 import { DesignSystemProvider } from "@orchest/design-system";
 import { domMax, LazyMotion } from "framer-motion";
 import React from "react";
-import { GlobalContextProvider } from "./GlobalContext";
+import { BrowserRouter as Router } from "react-router-dom";
+import { AppContextProvider } from "./AppContext";
+import { GlobalContextProvider, useGlobalContext } from "./GlobalContext";
 import { Intercom } from "./Intercom";
 import { ProjectsContextProvider } from "./ProjectsContext";
 import { SessionsContextProvider } from "./SessionsContext";
@@ -21,6 +24,14 @@ export const DesignProvider: React.FC = ({ children }) => {
   );
 };
 
+export const GlobalProviders: React.FC = ({ children }) => {
+  return (
+    <GlobalContextProvider>
+      <Intercom>{children}</Intercom>
+    </GlobalContextProvider>
+  );
+};
+
 /**
  * Providers that are specific to Orchest core, outer provider represents higher-level logic
  * - SessionsContext: handling sessions for the user
@@ -34,12 +45,51 @@ export const OrchestProvider: React.FC = ({ children }) => {
   );
 };
 
-export const GlobalProviders: React.FC = ({ children }) => {
+/**
+ * Sets up the router and a "unsaved changes" handler.
+ * Requires `GlobalContext`.
+ */
+export const RouteProvider: React.FC = ({ children }) => {
+  const { setAsSaved, setConfirm } = useGlobalContext();
+  const discardActiveCronJobChanges = useEditJob(
+    (state) => state.discardActiveCronJobChanges
+  );
+
+  return (
+    <Router
+      getUserConfirmation={(message, callback) => {
+        // use Prompt component to intercept route changes
+        // handle the blocking event here
+        if (message === "hasUnsavedChanges") {
+          setConfirm(
+            "Warning",
+            "There are unsaved changes. Are you sure you want to navigate away?",
+            async (resolve) => {
+              setAsSaved();
+              discardActiveCronJobChanges();
+              callback(true);
+              resolve(true);
+              return true;
+            }
+          );
+        }
+      }}
+    >
+      {children}
+    </Router>
+  );
+};
+
+export const AppProviders: React.FC = ({ children }) => {
   return (
     <DesignProvider>
-      <GlobalContextProvider>
-        <Intercom>{children}</Intercom>
-      </GlobalContextProvider>
+      <GlobalProviders>
+        <RouteProvider>
+          <OrchestProvider>
+            <AppContextProvider>{children}</AppContextProvider>
+          </OrchestProvider>
+        </RouteProvider>
+      </GlobalProviders>
     </DesignProvider>
   );
 };
